When the height is changed, the step should be the Propose step
In the past, since the Commit state is the first state of the height,
the step is changed to the Commit step after a block is imported.

However, in the current, the Commit step is the last step of the
height.

This patch fixes the unchanged code in the past.